+++ CHANGELOG.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `RadioButton` https://github.com/Textualize/textual/pull/1872
 - Added `RadioSet` https://github.com/Textualize/textual/pull/1872
 
+### Changed
+
+- Widget scrolling methods (such as `Widget.scroll_home` and `Widget.scroll_end`) now perform the scroll after the next refresh https://github.com/Textualize/textual/issues/1774
+
+### Fixed
+
+- Scrolling with cursor keys now moves just one cell https://github.com/Textualize/textual/issues/1897
+
 ### Fixed
 
 - Fix exceptions in watch methods being hidden on startup https://github.com/Textualize/textual/issues/1886